# Make `stim.Tableau.from_stabilizers` faster